Updated the mod for SMAPI 2.2
