+++ PetEnhancement/Action.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PetEnhancement
+namespace PetEnhancements
 {
     public enum Action
     {
